+++ packages/core/phase2/common/getOperationSequence/areAllResultsAlreadyPresentOnPnc/isMatchToPncAdjAndDis/index.ts
@@ -1,0 +1,1 @@
+export { default as isMatchToPncAdjAndDis } from "./isMatchToPncAdjAndDis"
