+++ packages/core/phase2/processPncUpdateDataset/processPhase2PncUpdateDataset.ts
@@ -5,7 +5,7 @@ import type { Trigger } from "../../phase1/types/Trigger"
 import Phase from "../../types/Phase"
 import type { PncUpdateDataset } from "../../types/PncUpdateDataset"
 import allPncOffencesContainResults from "../common/allPncOffencesContainResults"
-import getOperationSequence from "../common/getOperationSequence/getOperationSequence"
+import { getOperationSequence } from "../common/getOperationSequence"
 import type Phase2Result from "../types/Phase2Result"
 import { Phase2ResultType } from "../types/Phase2Result"
 import refreshOperationSequence from "./refreshOperationSequence"
